Add files via upload
A handyman android mobile application based on java, the application offer two kind of services, for the service providers to register and publish their services, experiences and price. in the other side the the service for clients to search filter and choose a handyman in the area where he is living.